Extract BeaconsGetServerSidePropsContext to middleware/
