+++ src/lib/handlePageRequest.ts
@@ -7,10 +7,8 @@ import {
   updateFormCache,
   withCookieRedirect,
 } from "./middleware";
-import {
-  BeaconsGetServerSidePropsContext,
-  withContainer,
-} from "./middleware/withContainer";
+import { BeaconsGetServerSidePropsContext } from "./middleware/BeaconsGetServerSidePropsContext";
+import { withContainer } from "./middleware/withContainer";
 import { Registration } from "./registration/registration";
 import { IRegistration } from "./registration/types";
 import { retrieveUserFormSubmissionId } from "./retrieveUserFormSubmissionId";
